# cdk-cdd-namecheck-addresscheck/infrastructure/swagger.yaml
openapi: 3.0.0
info:
  title: Document Analysis API
  version: 1.0.0
  description: API for document analysis service
paths:
...
  /analyses/{analysis_id}:
    get:
      summary: Get analysis by ID
      description: Returns detailed information about a specific analysis
      operationId: getAnalysis
      parameters:
        - name: analysis_id
          in: path
          description: Analysis ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis details
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - fetchedAt
                properties:
                  data:
                    $ref: '#/components/schemas/Analysis'
                  fetchedAt:
                    type: string
                    format: date-time
                    description: Timestamp when data was retrieved
        '404':
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/get-url:
    post:
      summary: Generate pre-signed URL for reading a document
      description: Generates a pre-signed URL valid for 60 seconds to read a document from S3
      operationId: getDocumentUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                objectKey:
                  type: string
                  description: The S3 object key of the document
              required:
                - objectKey
      responses:
        '200':
          description: Pre-signed URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: The pre-signed URL
                  expiresIn:
                    type: integer
                    description: URL expiration time in seconds
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
    AnalysisStatus:
      type: string
      enum:
        - PENDING
        - STARTED
        - PROCESSING
        - PROCESSED
        - ERROR
    DocumentType:
      type: string
      enum:
        - MIXED
        - BANK_STATEMENT
        - ANNUAL_REPORT
...
    Analysis:
      type: object
      required:
        - analysisId
        - description
        - documentType
        - objectsData
        - chatHistory
        - status
        - createdAt
        - lastUpdatedAt
        - yearMonth
      properties:
        analysisId:
          type: string
          format: uuid
        description:
          type: string
        documentType:
          $ref: '#/components/schemas/DocumentType'
        objectsData:
          type: array
          items:
            $ref: '#/components/schemas/ObjectData'
        chatHistory:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        yearMonth:
          type: string
          description: Year and month in YYYY-MM format
        createdAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        analysisParameters:
          type: object
          additionalProperties: true
          description: Optional parameters for the analysis
        analysisResults:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisResult'
    AnalysisResult:
      type: object
      required:
        - analysis
        - result
        - inputToken
        - outputToken
      properties:
        analysis:
          type: string
        result:
          type: string
        inputToken:
          type: integer
        outputToken:
          type: integer
    ObjectData:
      type: object
      required:
        - object
        - data
        - numberOfPages
        - tokenInput
        - tokenOutput
      properties:
        object:
          type: string
          description: S3 object key
        data:
          type: array
          items:
            $ref: '#/components/schemas/PageData'
        numberOfPages:
          type: integer
        tokenInput:
          type: integer
        tokenOutput:
          type: integer
    PageData:
      type: object
      required:
        - page
        - content
        - tokenInput
        - tokenOutput
      properties:
        page:
          type: integer
        content:
          type: string
        tokenInput:
          type: integer
        tokenOutput:
          type: integer
    ChatMessage:
      type: object
      required:
        - role
        - content
        - timestamp
      properties:
        role:
          type: string
          enum:
            - human
            - assistant
        content:
          type: string
        timestamp:
          type: string
          format: date-time 